test: update test

+++ integration_tests/select_rank_02.f90
@@ -2,9 +2,9 @@
 program select_rank_test
     implicit none
 
-    integer :: b(2) = [1,2]
-    integer :: c(2,2) = reshape([1,2,3,4],[2,2])
-    integer :: d(2,2,2) = reshape([1,2,3,4,5,6,7,8],[2,2,2])
+    real :: b(2) = [1,2]
+    real :: c(2,2) = reshape([1,2,3,4],[2,2])
+    real :: d(2,2,2) = reshape([1,2,3,4,5,6,7,8],[2,2,2])
 
     call check(b, 1)
     call check(c, 2)
@@ -13,7 +13,7 @@ program select_rank_test
 contains
 
     subroutine check(x, expected)
-        integer, intent(in) :: x(..)
+        real, intent(in) :: x(..)
         integer, intent(in) :: expected
 
         select rank(x)
